% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpprTotpmap.R
\name{fastaparser}
\alias{fastaparser}
\title{Extract protein description for TP-MAP compatability}
\usage{
fastaparser(filepathdb)
}
\arguments{
\item{filepathdb:}{Path to protein database (FASTA format)}
}
\value{
list with protein uniprot accession numbers and corresponding description line
}
\description{
Extract protein description for TP-MAP compatability
}
\examples{
fastaparser("C:/protein.fas")
}
