% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmitotppr.R
\name{tmtiheader_to_tpprheaders}
\alias{tmtiheader_to_tpprheaders}
\title{Translate TMTI headers to TPPR headers}
\usage{
tmtiheader_to_tpprheaders(directory_of_interest)
}
\arguments{
\item{directory_of_interest"}{Experiment folder inside FragPipe Output folder}
}
\value{
a list of a dictionary of old and new lables, as well as a list of temperature and TMT conversion
}
\description{
Translate TMTI headers to TPPR headers
}
\examples{
tmtiheader_to_tpprheaders("C:/FragPipeOutputfolder"/Treatment)
}
