% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmtiTotpmap_2D.R
\name{tmtiTotpmap_2D}
\alias{tmtiTotpmap_2D}
\title{Performs 2DTPP data file conversion FragPipe outputs to TP-MAP input}
\usage{
tmtiTotpmap_2D(resultPath, fastadatabsefile)
}
\arguments{
\item{resultPath:}{FragPipe output folder.}

\item{fastadatabsefile:}{.fas file of the protein database}
}
\value{
void
}
\description{
Performs 2DTPP data file conversion FragPipe outputs to TP-MAP input
}
\examples{
tmtiTotpmap_2D("c:/FragPipefolder", "c:/proteindatabase.fas")

}
