% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmitotppr.R
\name{tmtitotppr}
\alias{tmtitotppr}
\title{Convert FragPipe Output to TPP-R input (for 1DTPP analysis)}
\usage{
tmtitotppr(fragpipefolder, experimentlables)
}
\arguments{
\item{fragpipefolder:}{Path to the folder where FragPipe saves results}

\item{experimentlables:}{A vector of stringd to determine what labels are use for different experiment types}
}
\value{
Path to the created TPP-R configuration file
}
\description{
Convert FragPipe Output to TPP-R input (for 1DTPP analysis)
}
\examples{
tmtitpttpr("C:/FragPipeOutputfolder", c("Treatment", "Vehicle"))
}
