% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmtiTotppr_2D.R
\name{tmtitotppr_2D}
\alias{tmtitotppr_2D}
\title{Convert FragPipe Output to TPP-R input (for 2DTPP analysis)}
\usage{
tmtitotppr_2D(
  fragpipefolder,
  experimentlabels,
  concentrationlabels,
  compound_val
)
}
\arguments{
\item{fragpipefolder:}{Path to the folder where FragPipe saves results}

\item{experimentlables:}{A vector of stringd to determine what labels are use for different experiment types}
}
\value{
Path to the created TPP-R configuration file
}
\description{
Convert FragPipe Output to TPP-R input (for 2DTPP analysis)
}
\examples{
tmtitpttpr("C:/FragPipeOutputfolder", c("42_44","46_48","50_52","54_56","58_60","62_64"), c(0,0.005,0.05,0.5,2), c(2mg ATP))
}
