% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpprNormOneDTPP.R
\name{tpprNormOneDTPP}
\alias{tpprNormOneDTPP}
\title{Performs TPP-R 1DTPP Normalization in FragPipe outputs}
\usage{
tpprNormOneDTPP(configurepath, resultPath)
}
\arguments{
\item{configurepath:}{Path to configuration file.}

\item{resultPath:}{FragPipe output folder.}
}
\value{
void
}
\description{
Performs TPP-R 1DTPP Normalization in FragPipe outputs
}
\examples{
tpprNormOneDTPP("C:/FragPipeOutputfolder/TPP-TR_config.xlsx","C:/FragPipeOutputfolder")
}
