% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpprTotpmap.R
\name{tpprTotpmap}
\alias{tpprTotpmap}
\title{TPP-R output to TP-MAP input (for 1DTPP analysis)}
\usage{
tpprTotpmap(configurationfile_path, databasepath)
}
\arguments{
\item{configurationfile_path:}{Path to configuration file.}

\item{databasepath:}{Path to protein database (FASTA format)}
}
\value{
void
}
\description{
TPP-R output to TP-MAP input (for 1DTPP analysis)
}
\examples{
tpprTotpmap("C:/FragPipeOutputfolder/TPP-TR_config.xlsx","C:/protein.fas")
}
