% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoDconversion.R
\name{twoDConversion}
\alias{twoDConversion}
\title{twoDConversion: Function to convert protein.tsv filesinto TPPR input for 2DTPP analysis}
\usage{
twoDConversion(
  fragpipefolder,
  experimentlabels,
  concentrationlabels,
  compound_val
)
}
\arguments{
\item{fragpipefolder:}{Path to the folder where FragPipe saves results}

\item{experimentlables:}{A vector of strings to determine what labels are use for different experiment types}

\item{concentrationlabels:}{A vector of strings indicatin the temperatures used in each experiment (plex)}

\item{compound_val:}{A string indicating the compounf of interest (e.g. ATP)}
}
\value{
configsavepath: Path to the newly created configuration file
}
\description{
twoDConversion: Function to convert protein.tsv filesinto TPPR input for 2DTPP analysis
}
\examples{
twofragpipe <- "Z:/crojaram/TPP_Project/PXD012423/2DTPP/ATP_rep1/FP20-1_build23"
conc_labels <- c(0,0.005,0.05,0.5,2)
labels_exp <- c("42_44","46_48","50_52","54_56","58_60","62_64")
compound <-("ATP")
configtwo <- twoDConversion(twofragpipe,labels_exp,conc_labels,compound)
}
